+++ src/components/AlbumTrackList.tsx
@@ -2,22 +2,20 @@ import type { SubsonicSong } from '../api/subsonicTypes';
 import { songToTrack } from '../utils/songToTrack';
 import type { Track } from '../store/playerStoreTypes';
 import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { useTracklistColumns, type ColDef } from '../utils/useTracklistColumns';
+import { useTracklistColumns } from '../utils/useTracklistColumns';
 import { usePlayerStore } from '../store/playerStore';
 import { useTranslation } from 'react-i18next';
 import { useDragDrop } from '../contexts/DragDropContext';
 import { useIsMobile } from '../hooks/useIsMobile';
 import { useSelectionStore } from '../store/selectionStore';
 import {
-  CENTERED_COLS,
   COLUMNS,
-  isSortable,
-  type ColKey,
   type SortKey,
 } from '../utils/albumTrackListHelpers';
 import { TrackRow } from './albumTrackList/TrackRow';
 import { AlbumTrackListMobile } from './albumTrackList/AlbumTrackListMobile';
 import { TracklistColumnPicker } from './albumTrackList/TracklistColumnPicker';
+import { TracklistHeaderRow } from './albumTrackList/TracklistHeaderRow';
 
 export type { SortKey } from '../utils/albumTrackListHelpers';
 
@@ -162,103 +160,6 @@ export default function AlbumTrackList({
 
   const currentTrackId = currentTrack?.id ?? null;
 
-  const handleHeaderClick = (key: ColKey | string) => {
-    if (!isSortable(key) || !onSort) return;
-    onSort(key);
-  };
-
-  const renderSortIndicator = (key: SortKey) => {
-    if (sortKey !== key) return null;
-    return (
-      <span style={{ marginLeft: 4, fontSize: 10, opacity: 0.7 }}>
-        {sortDir === 'asc' ? '▲' : '▼'}
-      </span>
-    );
-  };
-
-  // ── Header cell renderer ──────────────────────────────────────────────────
-  const renderHeaderCell = (colDef: ColDef, colIndex: number) => {
-    const key = colDef.key as ColKey;
-    const isLastCol = colIndex === visibleCols.length - 1;
-    const isCentered = CENTERED_COLS.has(key);
-    const label = colDef.i18nKey ? t(`albumDetail.${colDef.i18nKey as string}`) : '';
-    const canSort = isSortable(key) && onSort;
-    const isActive = canSort && sortKey === key;
-
-    if (key === 'num') {
-      return (
-        <div key={key} className="track-num">
-          <span
-            className={`bulk-check${allSelected ? ' checked' : ''}${inSelectMode ? ' bulk-check-visible' : ''}`}
-            onClick={e => { e.stopPropagation(); toggleAll(); }}
-            style={{ cursor: 'pointer' }}
-          />
-          <span className="track-num-number">#</span>
-        </div>
-      );
-    }
-
-    if (key === 'title') {
-      const hasNextCol = colIndex + 1 < visibleCols.length;
-      return (
-        <div
-          key={key}
-          style={{
-            position: 'relative',
-            padding: 0,
-            margin: 0,
-            minWidth: 0,
-            overflow: 'hidden',
-            cursor: canSort ? 'pointer' : 'default',
-            userSelect: 'none',
-          }}
-          onClick={() => handleHeaderClick(key)}
-          className={isActive ? 'tracklist-header-cell-active' : ''}
-        >
-          <div style={{ display: 'flex', width: '100%', height: '100%', alignItems: 'center', justifyContent: 'flex-start', paddingLeft: 12 }}>
-            <span style={{ overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap', fontWeight: isActive ? 600 : 400 }}>{label}</span>
-            {canSort && renderSortIndicator(key as SortKey)}
-          </div>
-          {hasNextCol && (
-            <div className="col-resize-handle" onMouseDown={e => startResize(e, colIndex + 1, -1)} />
-          )}
-        </div>
-      );
-    }
-
-    const isResizable = !isLastCol;
-    return (
-      <div
-        key={key}
-        style={{
-          position: 'relative',
-          padding: 0,
-          margin: 0,
-          minWidth: 0,
-          overflow: 'hidden',
-          cursor: canSort ? 'pointer' : 'default',
-          userSelect: 'none',
-        }}
-        onClick={() => handleHeaderClick(key)}
-        className={isActive ? 'tracklist-header-cell-active' : ''}
-      >
-        <div
-          style={{
-            display: 'flex', width: '100%', height: '100%', alignItems: 'center',
-            justifyContent: isCentered ? 'center' : 'flex-start',
-            paddingLeft: isCentered ? 0 : 12,
-          }}
-        >
-          <span style={{ whiteSpace: 'nowrap', overflow: 'hidden', textOverflow: 'ellipsis', fontWeight: isActive ? 600 : 400 }}>{label}</span>
-          {canSort && isSortable(key) && renderSortIndicator(key as SortKey)}
-        </div>
-        {isResizable && (
-          <div className="col-resize-handle" onMouseDown={e => startResize(e, colIndex, 1)} />
-        )}
-      </div>
-    );
-  };
-
   if (isMobile) {
     return (
       <AlbumTrackListMobile
@@ -296,12 +197,18 @@ export default function AlbumTrackList({
         }}
       >
 
-      {/* ── Header ── */}
-      <div className="tracklist-header-wrapper">
-        <div className="tracklist-header" style={gridStyle}>
-          {visibleCols.map((colDef, colIndex) => renderHeaderCell(colDef, colIndex))}
-        </div>
-      </div>
+      <TracklistHeaderRow
+        visibleCols={visibleCols}
+        gridStyle={gridStyle}
+        sortKey={sortKey}
+        sortDir={sortDir}
+        onSort={onSort}
+        allSelected={allSelected}
+        inSelectMode={inSelectMode}
+        toggleAll={toggleAll}
+        startResize={startResize}
+        t={t}
+      />
 
       {/* ── Tracks ── */}
       {discNums.map(discNum => (
